Refactor spacial components to geometry
