+++ foundry/core/sprites/SpriteGroup.py
@@ -3,10 +3,9 @@ from typing import Protocol
 from attr import attrs
 from PySide6.QtGui import QColor, QImage, QPainter
 
+from foundry.core.geometry import Point, Size
 from foundry.core.graphics_set.GraphicsSet import GraphicsSetProtocol
 from foundry.core.palette.PaletteGroup import MutablePaletteGroupProtocol
-from foundry.core.point.Point import Point
-from foundry.core.size.Size import Size
 from foundry.core.sprites import SPRITE_SIZE
 from foundry.core.sprites.Sprite import SpriteProtocol
 from foundry.game.gfx.drawable import MASK_COLOR
