+++ foundry/core/sprites/__init__.py
@@ -1,3 +1,3 @@
-from foundry.core.size.Size import Size
+from foundry.core.geometry import Size
 
 SPRITE_SIZE: Size = Size(8, 16)
