+++ foundry/game/gfx/objects/EnemyItem.py
@@ -1,10 +1,10 @@
 from PySide6.QtCore import QRect, QSize
 from PySide6.QtGui import QColor, QImage, QPainter, Qt
 
+from foundry.core.geometry import Point
 from foundry.core.graphics_page.GraphicsPage import GraphicsPage
 from foundry.core.graphics_set.GraphicsSet import GraphicsSet, GraphicsSetProtocol
 from foundry.core.palette.PaletteGroup import MutablePaletteGroup
-from foundry.core.point.Point import Point
 from foundry.game.EnemyDefinitions import (
     EnemyDefinition,
     GeneratorType,
